增加 FlapDelegation 负责代理 Adapter 部分方法
