+++ flap/src/main/java/me/yifeiyuan/flap/FlapAdapter.kt
@@ -8,20 +8,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.LifecycleOwner
 import androidx.recyclerview.widget.RecyclerView
-import me.yifeiyuan.flap.delegate.AdapterDelegate
-import me.yifeiyuan.flap.delegate.AdapterDelegateManager
 import me.yifeiyuan.flap.delegate.IAdapterDelegateManager
 import me.yifeiyuan.flap.event.Event
 import me.yifeiyuan.flap.event.EventObserver
 import me.yifeiyuan.flap.event.EventObserverWrapper
 import me.yifeiyuan.flap.ext.*
-import me.yifeiyuan.flap.hook.AdapterHookManager
 import me.yifeiyuan.flap.hook.IAdapterHookManager
 import me.yifeiyuan.flap.hook.PreloadHook
 import me.yifeiyuan.flap.pool.ComponentPool
 import java.util.*
 import me.yifeiyuan.flap.service.IAdapterServiceManager
-import me.yifeiyuan.flap.service.AdapterServiceManager
 
 /**
  * FlapAdapter is a flexible and powerful Adapter that makes you enjoy developing with RecyclerView.
@@ -34,7 +30,7 @@ import me.yifeiyuan.flap.service.AdapterServiceManager
  * @since 2020/9/22
  * @since 3.0.0
  */
-open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManager by AdapterHookManager(), IAdapterDelegateManager by AdapterDelegateManager(), IAdapterServiceManager by AdapterServiceManager(), SwipeDragHelper.Callback {
+open class FlapAdapter(val delegation: FlapDelegation = FlapDelegation()) : RecyclerView.Adapter<Component<*>>(), IAdapterHookManager by delegation, IAdapterDelegateManager by delegation, IAdapterServiceManager by delegation, SwipeDragHelper.Callback {
 
     companion object {
         private const val TAG = "FlapAdapter"
@@ -64,17 +60,9 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
     private var useGlobalComponentPool = false
 
     /**
-     * 默认 AdapterDelegate，兜底处理
-     */
-    private var defaultAdapterDelegate: AdapterDelegate<*, *>? = null
-
-    /**
      * RecyclerView 滑动到底部触发预加载
      */
     private var preloadHook: PreloadHook? = null
-
-    private val viewTypeDelegateCache: MutableMap<Int, AdapterDelegate<*, *>?> = mutableMapOf()
-    private val delegateViewTypeCache: MutableMap<AdapterDelegate<*, *>, Int> = mutableMapOf()
 
     /**
      * 所有事件的监听
@@ -108,7 +96,7 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
         adapterDelegates.addAll(Flap.adapterDelegates)
         adapterServices.putAll(Flap.adapterServices)
 
-        defaultAdapterDelegate = Flap.globalDefaultAdapterDelegate
+        delegation.fallbackDelegate = Flap.globalDefaultAdapterDelegate
 
         inflateWithApplicationContext = Flap.inflateWithApplicationContext
     }
@@ -146,29 +134,8 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Component<*> {
-        val delegate = getDelegateByViewType(viewType)
-        dispatchOnCreateViewHolderStart(this, delegate, viewType)
         val layoutInflater = LayoutInflater.from(if (inflateWithApplicationContext) parent.context.applicationContext else parent.context)
-        val component = delegate.onCreateViewHolder(layoutInflater, parent, viewType)
-        dispatchOnCreateViewHolderEnd(this, delegate, viewType, component)
-        return component
-    }
-
-    private fun dispatchOnCreateViewHolderStart(adapter: FlapAdapter, delegate: AdapterDelegate<*, *>, viewType: Int) {
-        adapterHooks.forEach {
-            it.onCreateViewHolderStart(adapter, delegate, viewType)
-        }
-    }
-
-    private fun dispatchOnCreateViewHolderEnd(
-            adapter: FlapAdapter,
-            delegate: AdapterDelegate<*, *>,
-            viewType: Int,
-            component: Component<*>
-    ) {
-        adapterHooks.forEach {
-            it.onCreateViewHolderEnd(adapter, delegate, viewType, component)
-        }
+        return delegation.onCreateViewHolder(this, parent, viewType, layoutInflater)
     }
 
     override fun onBindViewHolder(component: Component<*>, position: Int) {
@@ -180,49 +147,9 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
             position: Int,
             payloads: MutableList<Any>
     ) {
-        try {
-            val delegate = getDelegateByViewType(component.itemViewType)
-            val data = getItemData(position)
-            dispatchOnBindViewHolderStart(this, delegate, component, data, position, payloads)
-            delegate.onBindViewHolder(
-                    component,
-                    data,
-                    position,
-                    payloads,
-                    this
-            )
-            dispatchOnBindViewHolderEnd(this, delegate, component, data, position, payloads)
-            tryAttachLifecycleOwner(component)
-        } catch (e: Exception) {
-            e.printStackTrace()
-            FlapDebug.e(TAG, "onBindViewHolder: Error = ", e)
-        }
-    }
-
-    private fun dispatchOnBindViewHolderStart(
-            adapter: FlapAdapter,
-            delegate: AdapterDelegate<*, *>,
-            component: Component<*>,
-            data: Any,
-            position: Int,
-            payloads: MutableList<Any>
-    ) {
-        adapterHooks.forEach {
-            it.onBindViewHolderStart(adapter, delegate, component, data, position, payloads)
-        }
-    }
-
-    private fun dispatchOnBindViewHolderEnd(
-            adapter: FlapAdapter,
-            delegate: AdapterDelegate<*, *>,
-            component: Component<*>,
-            data: Any,
-            position: Int,
-            payloads: MutableList<Any>
-    ) {
-        adapterHooks.forEach {
-            it.onBindViewHolderEnd(adapter, delegate, component, data, position, payloads)
-        }
+        val data = getItemData(position)
+        delegation.onBindViewHolder(this, data, component, position, payloads)
+        tryAttachLifecycleOwner(component)
     }
 
     open fun getItemData(position: Int): Any {
@@ -230,46 +157,13 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
     }
 
     override fun getItemViewType(position: Int): Int {
-        var itemViewType: Int
         val itemData = getItemData(position)
-
-        val delegate: AdapterDelegate<*, *>? = adapterDelegates.firstOrNull {
-            it.delegate(itemData)
-        } ?: defaultAdapterDelegate
-
-        if (delegateViewTypeCache.containsKey(delegate)) {
-            return delegateViewTypeCache[delegate]!!
-        }
-
-        itemViewType = delegate?.getItemViewType(itemData)
-                ?: throw AdapterDelegateNotFoundException("找不到对应的 AdapterDelegate，请先注册或设置默认 AdapterDelegate ，position=$position , itemData=$itemData")
-
-        if (itemViewType == 0) {
-            itemViewType = generateItemViewType()
-        }
-        FlapDebug.d(TAG, "getItemViewType() called with: position = $position , itemViewType = $itemViewType")
-        viewTypeDelegateCache[itemViewType] = delegate
-        delegateViewTypeCache[delegate] = itemViewType
-        return itemViewType
-    }
-
-    private fun generateItemViewType(): Int {
-        val viewType = ViewTypeGenerator.generateViewType()
-        if (viewTypeDelegateCache.containsKey(viewType)) {
-            return generateItemViewType()
-        }
-        return viewType
-    }
-
-    fun getDelegateByViewType(viewType: Int): AdapterDelegate<*, *> {
-        return viewTypeDelegateCache[viewType] ?: defaultAdapterDelegate
-        ?: throw AdapterDelegateNotFoundException("找不到 viewType = $viewType 对应的 Delegate，请先注册，或设置默认的 Delegate")
+        return delegation.getItemViewType(position, itemData)
     }
 
     override fun getItemId(position: Int): Long {
         val itemData = getItemData(position)
-        val delegate = getDelegateByViewType(getItemViewType(position))
-        return delegate.getItemId(itemData, position)
+        return delegation.getItemId(position, itemData)
     }
 
     /**
@@ -289,6 +183,7 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
     override fun onAttachedToRecyclerView(recyclerView: RecyclerView) {
         super.onAttachedToRecyclerView(recyclerView)
         handleOnAttachedToRecyclerView(recyclerView)
+        delegation.onAttachedToRecyclerView(this, recyclerView)
     }
 
     private fun handleOnAttachedToRecyclerView(recyclerView: RecyclerView) {
@@ -323,35 +218,26 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
         if (this::componentPool.isInitialized) {
             bindingContext.applicationContext.unregisterComponentCallbacks(componentPool)
         }
+        delegation.onDetachedFromRecyclerView(this, recyclerView)
     }
 
     /**
      * 会优先于 FlapComponentPool.putRecycledView 被调用
      */
     override fun onViewRecycled(component: Component<*>) {
-        val delegate = getDelegateByViewType(component.itemViewType)
-        delegate.onViewRecycled(this, component)
+        delegation.onViewRecycled(this, component)
     }
 
     override fun onFailedToRecycleView(component: Component<*>): Boolean {
-        val delegate = getDelegateByViewType(component.itemViewType)
-        return delegate.onFailedToRecycleView(this, component)
+        return delegation.onFailedToRecycleView(this, component)
     }
 
     override fun onViewAttachedToWindow(component: Component<*>) {
-        val delegate = getDelegateByViewType(component.itemViewType)
-        delegate.onViewAttachedToWindow(this, component)
-        adapterHooks.forEach {
-            it.onViewAttachedToWindow(this, delegate, component)
-        }
+        delegation.onViewAttachedToWindow(this, component)
     }
 
     override fun onViewDetachedFromWindow(component: Component<*>) {
-        val delegate = getDelegateByViewType(component.itemViewType)
-        delegate.onViewDetachedFromWindow(this, component)
-        adapterHooks.forEach {
-            it.onViewDetachedFromWindow(this, delegate, component)
-        }
+        delegation.onViewDetachedFromWindow(this, component)
     }
 
     /**
@@ -521,9 +407,3 @@ open class FlapAdapter : RecyclerView.Adapter<Component<*>>(), IAdapterHookManag
         swapData(fromPosition, toPosition)
     }
 }
-
-/**
- * 当 Adapter.data 中存在一个 Model 没有对应的 AdapterDelegate.delegate()==true 时抛出
- */
-internal class AdapterDelegateNotFoundException(errorMessage: String) : Exception(errorMessage)
-
